feat: Move MailDateFormatUtils to the new date package inside utils
