+++ app/src/main/java/com/infomaniak/mail/utils/PrintHeaderUtils.kt
@@ -1,6 +1,6 @@
 /*
  * Infomaniak Mail - Android
- * Copyright (C) 2024 Infomaniak Network SA
+ * Copyright (C) 2024-2025 Infomaniak Network SA
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -21,7 +21,7 @@ import android.content.Context
 import com.infomaniak.mail.R
 import com.infomaniak.mail.data.models.correspondent.Recipient
 import com.infomaniak.mail.data.models.message.Message
-import com.infomaniak.mail.utils.MailDateFormatUtils.formatForHeader
+import com.infomaniak.mail.utils.date.MailDateFormatUtils.formatForHeader
 import com.infomaniak.mail.utils.extensions.formatSubject
 import com.infomaniak.mail.utils.extensions.toDate
 import org.jsoup.nodes.Element
