+++ app/src/main/java/com/infomaniak/mail/utils/date/MailDateFormatUtils.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.utils
+package com.infomaniak.mail.utils.date
 
 import android.content.Context
 import android.os.Build
